// src/app/features/basket/models/status.ts
export enum Status {
  STARTED_MAKING = 'STARTED_MAKING',
  WAITING_FOR_DELIVERYMAN = 'WAITING_FOR_DELIVERYMAN',
  TO_RESTAURANT= 'TO_RESTAURANT',
  GET_THE_ORDER = 'GET_THE_ORDER',
  ON_THE_WAY_TO_CUSTOMER = 'ON_THE_WAY_TO_CUSTOMER',
  WAITING_FOR_CUSTOMER = 'WAITING_FOR_CUSTOMER',
  READY = 'READY',
}